feat: automatic download from source

+++ helper/read_chunk.go
@@ -1,6 +1,7 @@
 package helper
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/canhlinh/hlsdl"
 	"io"
@@ -8,38 +9,81 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
+	"strconv"
 	"strings"
+	"sync"
+	"time"
 )
 
-func ReadChunk(url, filename string) {
-	resp, err := http.Get(url + "/playlist.m3u8")
-	if err != nil {
-		log.Println(err)
+type Source struct {
+	Name string `json:"name"`
+	Uri  string `json:"uri"`
+}
+
+func ReadChunk() {
+	var wg sync.WaitGroup
+	for {
+		sources := ReadJSON()
+
+		for i := 0; i < len(sources); i++ {
+			wg.Add(1)
+			go processSource(sources[i], &wg)
+		}
+
+		wg.Wait()
+
+		time.Sleep(3 * time.Second)
 	}
+}
+
+func processSource(source Source, wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	fmt.Println("Downloading: " + source.Name)
+	resp, err := http.Get(source.Uri + "/playlist.m3u8")
+	if err != nil {
+		log.Println("Error downloading from", source.Name, ":", err)
+		return
+	}
+	defer resp.Body.Close()
 
 	bytes, err := io.ReadAll(resp.Body)
 	if err != nil {
-		log.Println(err)
+		log.Println("Error reading response from", source.Name, ":", err)
+		return
 	}
 
 	stringResponse := string(bytes)
 
 	splitResponse := strings.Split(stringResponse, "\n")
 
-	chunk := splitResponse[len(splitResponse)-2]
+	if len(splitResponse) >= 2 {
+		chunk := splitResponse[len(splitResponse)-2]
 
-	download, err := hlsdl.New(url+chunk, nil, "download", "", 8, false).Download()
-	if err != nil {
-		return
-	}
+		fmt.Println("Response: ", len(splitResponse))
 
-	fmt.Println("Downloaded: " + download)
+		download, err := hlsdl.New(source.Uri+chunk, nil, "download", "", 8, false).Download()
+		if err != nil {
+			log.Println("Error downloading chunk from", source.Name, ":", err)
+			return
+		}
 
-	ConvertToMp4(download, "./result/"+filename)
+		fmt.Println("Downloaded: " + download)
 
-	err = os.Remove(download)
-	if err != nil {
-		fmt.Println(err)
+		ConvertToMp4(download, "result/"+source.Name+"/"+source.Name+strconv.Itoa(int(time.Now().Unix()))+".mp4")
+
+		time.Sleep(2 * time.Second)
+
+		if _, err := os.Stat(download); err == nil {
+			err = os.Remove(download)
+			if err != nil {
+				fmt.Println("Error removing downloaded chunk from", source.Name, ":", err)
+			}
+		} else {
+			fmt.Println("File not found:", download)
+		}
+	} else {
+		log.Println("Not enough elements in splitResponse for", source.Name)
 	}
 }
 
@@ -54,5 +98,45 @@ func ConvertToMp4(filenameIn, filenameOut string) {
 	}
 
 	fmt.Println("Conversion successful!")
+}
 
+func ReadJSON() []Source {
+	open, err := os.Open("./helper/source.json")
+	if err != nil {
+		fmt.Println("Error opening source.json:", err)
+		return nil
+	}
+	defer func(open *os.File) {
+		err := open.Close()
+		if err != nil {
+			fmt.Println("Error closing source.json:", err)
+		}
+	}(open)
+
+	bytes, err := io.ReadAll(open)
+	if err != nil {
+		fmt.Println("Error reading source.json:", err)
+		return nil
+	}
+
+	var sources []Source
+
+	err = json.Unmarshal(bytes, &sources)
+	if err != nil {
+		fmt.Println("Error unmarshalling source.json:", err)
+		return nil
+	}
+
+	return sources
+}
+
+func InitFolder() {
+	dir := ReadJSON()
+
+	for i := 0; i < len(dir); i++ {
+		err := os.Mkdir("result/"+dir[i].Name, 0755)
+		if err != nil {
+			fmt.Println("Error creating directory for", dir[i].Name, ":", err)
+		}
+	}
 }
